+++ compiler-core/src/parse.rs
@@ -3006,7 +3006,6 @@ where
         let mut start = 0;
         let mut end;
         let mut module = EcoString::new();
-        let mut last_segment_start;
         let mut last_segment_end;
 
         // Gather module names
@@ -3019,7 +3018,6 @@ where
             }
             module.push_str(&name);
             end = e;
-            last_segment_start = s;
             last_segment_end = e;
 
             // Useful error for : import a/.{b}
@@ -3100,7 +3098,7 @@ where
                 end,
             },
             module_location: SrcSpan {
-                start: last_segment_start,
+                start,
                 end: last_segment_end,
             },
             unqualified_values,